+++ changelog.md
@@ -2,7 +2,8 @@
 1. Infinity Armor received a buff in speed
 2. All Infinity tools now have +INFINITY damage
 3. ACHIEVEMENTS
-4. Infinity Banana now has 50% chance kill you (without armor)
+4. Infinity Banana now kill you (without armor)
+5. Infinity Bow kill you if not wearing full Infinity Armor
 
 # 1.2
 1. Storage Drawers compat